moved the motor diagnostics to the diagnostics_specific.hpp file, keeping motor_test files for running motor tests: PWM offset / strain gauge etc...
