+++ sr_self_test/include/sr_self_test/diagnostics_parser/diagnostics_specific.hpp
@@ -38,7 +38,6 @@
 #include <ros/ros.h>
 
 #include "sr_self_test/diagnostics_parser/diagnostics_common.hpp"
-#include "sr_self_test/diagnostics_parser/motor_test.hpp"
 
 namespace shadow_robot
 {
@@ -89,6 +88,33 @@ namespace shadow_robot
     virtual std::auto_ptr<BaseDiagnostics> shallow_clone(std::string name)
     {
       std::auto_ptr<BaseDiagnostics> tmp( new EtherCATMasterDiagnostics(name, test_runner_) );
+      return tmp;
+    };
+  };
+
+
+  class MotorDiagnostics
+    : public MinMaxDiagnostics
+  {
+  public:
+    MotorDiagnostics(std::string name, self_test::TestRunner* test_runner)
+      : MinMaxDiagnostics(name, test_runner)
+    {
+      values_.reset(new DiagMap() );
+      DiagnosticTest voltage;
+      voltage.min_max = std::make_pair(23.5, 24.5); //min and max acceptable voltage
+      values_->insert( std::pair<std::string, DiagnosticTest>("Measured Voltage", voltage) );
+
+      DiagnosticTest temperature;
+      temperature.min_max = std::make_pair(20.0, 50.0); //min and max acceptable temperature
+      values_->insert( std::pair<std::string, DiagnosticTest>("Temperature", temperature) );
+    };
+
+    ~MotorDiagnostics() {};
+
+    virtual std::auto_ptr<BaseDiagnostics> shallow_clone(std::string name)
+    {
+      std::auto_ptr<BaseDiagnostics> tmp( new MotorDiagnostics(name, test_runner_) );
       return tmp;
     };
   };
