+++ sr_self_test/src/motor_test.cpp
@@ -19,37 +19,16 @@
  * with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  *
- * @brief Run different tests on the motors (parses diagnostics, apply PWM offset
+ * @brief Run different tests on the motors (apply PWM offset
  *        and measure current, test strain gauges, etc.. )
  *
  *
  */
 
-#include "sr_self_test/diagnostics_parser/motor_test.hpp"
+#include "sr_self_test/motor_test.hpp"
 
 namespace shadow_robot
 {
-  MotorDiagnostics::MotorDiagnostics(std::string name, self_test::TestRunner* test_runner)
-    : MinMaxDiagnostics(name, test_runner)
-  {
-    values_.reset(new DiagMap() );
-    DiagnosticTest voltage;
-    voltage.min_max = std::make_pair(23.5, 24.5); //min and max acceptable voltage
-    values_->insert( std::pair<std::string, DiagnosticTest>("Measured Voltage", voltage) );
-
-    DiagnosticTest temperature;
-    temperature.min_max = std::make_pair(20.0, 50.0); //min and max acceptable temperature
-    values_->insert( std::pair<std::string, DiagnosticTest>("Temperature", temperature) );
-  }
-
-  MotorDiagnostics::~MotorDiagnostics()
-  {};
-
-  std::auto_ptr<BaseDiagnostics> MotorDiagnostics::shallow_clone(std::string name)
-  {
-    std::auto_ptr<BaseDiagnostics> tmp( new MotorDiagnostics(name, test_runner_) );
-    return tmp;
-  };
 }
 
 /* For the emacs weenies in the crowd.
